requirements.txt
# Core ML Frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.21.0
accelerate>=0.20.0

# Computer Vision
Pillow>=9.5.0
opencv-python>=4.8.0
numpy>=1.24.0,<2.0.0

# Web Framework
Flask>=2.3.0
flask-cors>=4.0.0

# Evaluation Metrics
scikit-image>=0.21.0
scipy>=1.11.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0